requirements.txt
anyio==4.6.2.post1
appdirs==1.4.4
certifi==2024.8.30
colorama==0.4.6
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
ollama==0.3.3
pyperclip==1.9.0
PyYAML==6.0.2
sniffio==1.3.1
